Move the store debugger to a new file.
